egldevice: implement eglQueryDisplayAttribEXT
This adds a new vfunc to _EGLDriver, QueryDeviceName, which should
return a const string of the device name (usually in the format
'/dev/dri/cardN').

The EGLDevice could perhaps be cached in the EGLDisplay but there
usually aren't loads of devices and this lookup isn't particularly
costly so leave it as is for now.

Right now this only works with the egl_dri2 driver.

Signed-off-by: Jonny Lamb <jonny.lamb@collabora.co.uk>